% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mse.R
\name{calc_mse}
\alias{calc_mse}
\title{Calculate Mean Squared Error}
\usage{
calc_mse(y, y_pred)
}
\arguments{
\item{y}{a numeric vector of actual response values}

\item{y_pred}{a numeric vector of predicted response values}
}
\value{
a numeric value of the mean squared error
}
\description{
Calculates the mean squared error of a linear regression model
}
\examples{
y <- c(2, 3, 5, 7)
y_pred <- c(1.5, 3.5, 4.5, 6.5)
mse <- calc_mse(y, y_pred)
mse # display the mean squared error

}
