% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_regression.R
\name{compare_regression_models}
\alias{compare_regression_models}
\title{Compare Regression Models}
\usage{
compare_regression_models(
  data,
  x_vars,
  y_var,
  new_data,
  lambda = 1,
  alpha = 1,
  max_iter = 1000,
  tol = 1e-04,
  degree = 2,
  k = 5
)
}
\arguments{
\item{data}{A data frame containing the data to fit the models}

\item{x_vars}{A character vector of predictor variable names}

\item{y_var}{A character string of the response variable name}

\item{new_data}{A data frame containing the data for making predictions}

\item{lambda}{A numeric value for the ridge and lasso regularization parameter (default 1)}

\item{alpha}{A numeric value for the lasso mixing parameter (default 1)}

\item{max_iter}{A numeric value for the maximum number of iterations for lasso (default 1000)}

\item{tol}{A numeric value for the tolerance for lasso convergence (default 1e-4)}

\item{degree}{A numeric value for the polynomial regression degree (default 2)}

\item{k}{A numeric value for the number of neighbors for kNN regression (default 5)}
}
\value{
A data frame showing the models and their corresponding MSE values, sorted by smallest MSE
}
\description{
This function compares different regression models by calculating their mean squared error (MSE)
}
