% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_linear_regression.R
\name{fit_linear_regression}
\alias{fit_linear_regression}
\title{Fit Linear Regression Model}
\usage{
fit_linear_regression(data, x_vars, y_var, intercept = TRUE)
}
\arguments{
\item{data}{dataframe containing the feature and label variables}

\item{x_vars}{character vector of variable names to be used as predictors}

\item{y_var}{character string of the variable name to be used as the response variable}

\item{intercept}{logical value indicating whether to include an intercept term in the model}
}
\value{
a list containing the model object and fitted values
}
\description{
Fits a linear regression model to the given data
}
\examples{
data(mtcars)
lm_fit <- fit_linear_regression(data = mtcars, x_vars = c("wt"), y_var = "mpg")
lm_fit$model # display the model object
lm_fit$fitted_values # display the fitted values

}
