% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ridge_regression.R
\name{fit_ridge_regression}
\alias{fit_ridge_regression}
\title{Fit Ridge Regression Model}
\usage{
fit_ridge_regression(data, x_vars, y_var, lambda, intercept = TRUE)
}
\arguments{
\item{data}{dataframe containing the feature and label variables}

\item{x_vars}{character vector of variable names to be used as predictors}

\item{y_var}{character string of the variable name to be used as the response variable}

\item{lambda}{the value of the regularization parameter}

\item{intercept}{logical value indicating whether to include an intercept term in the model}
}
\value{
a list containing the model object and fitted values
}
\description{
Fits a Ridge regression model to the given data
}
\examples{
data(mtcars)
ridge_fit <- fit_ridge_regression(data = mtcars, x_vars = c("wt"), y_var = "mpg", lambda = 0.1)
ridge_fit$model # display the model object
ridge_fit$fitted_values # display the fitted values

}
