% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation_matrix.R
\name{plot_correlation_matrix}
\alias{plot_correlation_matrix}
\title{Plot Correlation Matrix}
\usage{
plot_correlation_matrix(
  data,
  title = "Correlation Matrix",
  size_coercoef = 0.5
)
}
\arguments{
\item{data}{A data frame containing numeric variables}

\item{title}{A character string representing the title of the plot (optional)}

\item{size_coercoef}{A numeric representing the size for the correlation coefficients}
}
\value{
A corrplot object displaying the correlation matrix plot, or an error message if an exception occurs
}
\description{
This function creates a correlation matrix plot of all variables in a given dataset using the corrplot package
}
