% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_y_actual_vs_predicted.R
\name{plot_y_actual_vs_predicted}
\alias{plot_y_actual_vs_predicted}
\title{Plot Actual vs. Predicted Values}
\usage{
plot_y_actual_vs_predicted(y_actual, y_pred, model_name)
}
\arguments{
\item{y_actual}{A numeric vector of the actual response variable values}

\item{y_pred}{A numeric vector of the predicted response variable values}

\item{model_name}{A character string representing the name of the regression model used for predictions}
}
\value{
A ggplot object displaying the scatterplot of actual vs. predicted values, or an error message if an exception occurs
}
\description{
This function creates a scatterplot of actual vs. predicted values for a given regression model
}
