% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_KNN_regression.R
\name{predict_knn_regression}
\alias{predict_knn_regression}
\title{Predict using the KNN Regression Model}
\usage{
predict_knn_regression(data, x_vars, y_var, new_data, k = 5)
}
\arguments{
\item{data}{A data frame containing the predictor and response variables}

\item{x_vars}{A character vector of column names corresponding to the predictor variables}

\item{y_var}{A character string of the column name corresponding to the response variable}

\item{new_data}{A data frame containing the new predictor variable values}

\item{k}{The number of nearest neighbors to consider (default = 5)}
}
\value{
A vector of predicted response variable values
}
\description{
This function predicts the response variable using the KNN Regression model.
}
\examples{
# Generate sample data
set.seed(123)
n <- 100
p <- 2
data <- data.frame(matrix(runif(n * p, -1, 1), n, p))
colnames(data) <- paste0("x", 1:p)
data$y <- data$x1 + data$x2 + rnorm(n, sd = 0.1)

# Generate new data
new_data <- data.frame(matrix(runif(10 * p, -1, 1), 10, p))
colnames(new_data) <- paste0("x", 1:p)

# Predict the response variable using the KNN model
predictions <- predict_knn_regression(data, x_vars=paste0("x", 1:p), y_var="y", new_data, k=5)
print(predictions)
}
