% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_linear_regression.R
\name{predict_linear_regression}
\alias{predict_linear_regression}
\title{Predict Using Linear Regression Model}
\usage{
predict_linear_regression(data, lm_fit, x_vars, intercept = TRUE)
}
\arguments{
\item{data}{dataframe containing the feature variables}

\item{lm_fit}{a list containing the model object and fitted values}

\item{x_vars}{character vector of variable names to be used as predictors}

\item{intercept}{logical value indicating whether to include an intercept term in the model}
}
\value{
a numeric vector of predicted values
}
\description{
Predicts the response variable using a linear regression model
}
\examples{
data(mtcars)
lm_fit <- fit_linear_regression(data = mtcars, x_vars = c("wt"), y_var = "mpg")
lm_pred <- predict_linear_regression(data = mtcars, lm_fit, x_vars = c("wt"))
lm_pred # display the predicted values

}
