% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ridge_regression.R
\name{predict_ridge_regression}
\alias{predict_ridge_regression}
\title{Predict Using Ridge Regression Model}
\usage{
predict_ridge_regression(data, ridge_fit, x_vars)
}
\arguments{
\item{data}{dataframe containing the feature variables}

\item{ridge_fit}{a list containing the model object and fitted values}

\item{x_vars}{character vector of variable names to be used as predictors}
}
\value{
a numeric vector of predicted values
}
\description{
Predicts the response variable using a Ridge regression model
}
\examples{
data(mtcars)
ridge_fit <- fit_ridge_regression(data = mtcars, x_vars = c("wt"), y_var = "mpg", lambda = 0.1)
ridge_pred <- predict_ridge_regression(data = mtcars, ridge_fit, x_vars = c("wt"))
ridge_pred # display the predicted values

}
